% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiUAPI.R
\name{distanfu}
\alias{distanfu}
\title{This function shows the distance between 2  places}
\usage{
distanfu(place1, place2, f)
}
\arguments{
\item{numeric}{or character , enter two city name or particular place like 2 adrresses.
and the key to the API}
}
\value{
it returns the distance betwen places by car.
}
\description{
This function shows the distance between 2  places
}
