\name{LiUAPI-package}
\alias{LiUAPI-package}
\alias{LiUAPI}
\docType{package}
\title{
\packageTitle{LiUAPI}
}
\description{
\packageDescription{LiUAPI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LiUAPI}
\packageIndices{LiUAPI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LiUAPI}

Maintainer: \packageMaintainer{LiUAPI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
