% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiUAPI.R
\name{places}
\alias{places}
\title{This function shows nearby places}
\usage{
places(input, f)
}
\arguments{
\item{numeric}{or character , enter a city name and a particular place like resturant or park.
and the key to the API}
}
\value{
it returns nearby places by detail.
}
\description{
This function shows nearby places
}
